Vendor xgrammar-0.1.26 source and implement missing APIs
- Vendor all xgrammar-0.1.26 C++ source code (11M) for crates.io compatibility
- Remove embedded git repos from 3rdparty dependencies (cpptrace, dlpack, googletest)
- Implement BatchGrammarMatcher for parallel batch processing
- Add reset_token_bitmask() utility function
- Add Grammar::from_structural_tag() method with error handling
- Reorganize matcher module: split into mod.rs, grammar_matcher.rs, batch_grammar_matcher.rs
- Add comprehensive C++ FFI helpers for batch operations
- API coverage improved from 70% to 78%
- All 87 tests passing

+++ src/compiler/compiled_grammar.rs
@@ -30,7 +30,9 @@ impl CompiledGrammar {
 
     /// The approximate memory usage of the compiled grammar in bytes.
     pub fn memory_size_bytes(&self) -> usize {
-        self.inner.MemorySizeBytes()
+        // MemorySizeBytes comes from C++ and maps to Rust's usize via cxx.
+        // Avoid usize -> u64 -> usize conversions which are not universally infallible.
+        self.inner.MemorySizeBytes() as usize
     }
 
     /// Serialize the compiled grammar to a JSON string.
